page-data.ts for default page.tsx products queries
