+++ core/app/[locale]/(default)/page-data.ts
@@ -1,0 +1,45 @@
+import { removeEdgesAndNodes } from '@bigcommerce/catalyst-client';
+import { getSessionCustomerId } from '~/auth';
+import { client } from '~/client';
+import { graphql } from '~/client/graphql';
+import { revalidate } from '~/client/revalidate-target';
+import { ProductCardCarouselFragment } from '~/components/product-card-carousel';
+
+const HomePageQuery = graphql(
+  `
+    query HomePageQuery {
+      site {
+        newestProducts(first: 12) {
+          edges {
+            node {
+              ...ProductCardCarouselFragment
+            }
+          }
+        }
+        featuredProducts(first: 12) {
+          edges {
+            node {
+              ...ProductCardCarouselFragment
+            }
+          }
+        }
+      }
+    }
+  `,
+  [ProductCardCarouselFragment],
+);
+
+export const getPageData = async () => {
+  const customerId = await getSessionCustomerId();
+
+  const { data } = await client.fetch({
+    document: HomePageQuery,
+    customerId,
+    fetchOptions: customerId ? { cache: 'no-store' } : { next: { revalidate } },
+  });
+
+  return {
+    featuredProducts: removeEdgesAndNodes(data.site.featuredProducts),
+    newestProducts: removeEdgesAndNodes(data.site.newestProducts),
+  };
+};
